Wrap into Flask
