+++ .env_dev
@@ -3,3 +3,4 @@ SAMPLE_EMAIL=/home/atmos/Hobby/FlaskExploration/data/emails/emailSample1.txt
 MODEL=/home/atmos/Hobby/FlaskExploration/models/svm_v_0_0_1.pkl
 TRAIN_DATA=/home/atmos/Hobby/FlaskExploration/data/TrainMatrix.csv
 TEST_DATA=/home/atmos/Hobby/FlaskExploration/data/TestMatrix.csv
+DO_MODEL_UPDATE=true
