+++ models/svm_v_0_0_1_report.csv
@@ -1,2 +1,2 @@
 model,best_score,C,kernel,gamma,AUC,time_elapsed
-SVC,0.9962773335610047,100,rbf,0.001,0.9973772614668569,48.01316022872925
+SVC,0.9962773335610047,100,rbf,0.001,0.9973772614668569,46.166532039642334
